code/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
flask-restful
flask-security-too
Flask-Caching
numpy
matplotlib
requests
redis
pandas